deploy/requirements.txt
requests>=2.32
python-dotenv>=1.0
fastapi>=0.110
uvicorn>=0.29
botbuilder-core>=4.17.1
botbuilder-integration-aiohttp>=4.17.1
aiofiles>=23.0
